+++ Sources/MapboxMaps/Foundation/Extensions/Core/OfflineRegion.swift
@@ -1,0 +1,26 @@
+@_implementationOnly import MapboxCoreMaps_Private
+
+@available(*, deprecated)
+extension OfflineRegion {
+    // :nodoc:
+    func invalidate(completion: @escaping (Result<Void, Error>) -> Void) {
+        self.invalidate(forCallback: coreAPIClosureAdapter(for: { error in
+            if let error = error {
+                completion(.failure(error))
+            } else {
+                completion(.success(()))
+            }
+        }, concreteErrorType: MapError.self))
+    }
+
+    // :nodoc:
+    func purge(completion: @escaping (Result<Void, Error>) -> Void) {
+        self.purge(forCallback: coreAPIClosureAdapter(for: { error in
+            if let error = error {
+                completion(.failure(error))
+            } else {
+                completion(.success(()))
+            }
+        }, concreteErrorType: MapError.self))
+    }
+}